client_slack/requirements.txt
slack_bolt
mcp
openai
python-dotenv
pandas
aiohttp
pypdf
python-docx
